+++ config.php
@@ -170,6 +170,7 @@ $THEME->layouts = array(
 
 $THEME->rendererfactory = 'theme_overridden_renderer_factory';
 
+$THEME->prescsscallback = 'theme_foundation_pre_scss';
 $THEME->scss = function(theme_config $theme) {
     $toolbox = \theme_foundation\toolbox::get_instance();
     $scss = $toolbox->get_main_scss_content($theme);