+++ lib.php
@@ -58,6 +58,17 @@ function theme_foundation_pluginfile($course, $cm, $context, $filearea, $args, $
 }
 
 /**
+ * Gets the pre SCSS for the theme.
+ *
+ * @param theme_config $theme The theme configuration object.
+ * @return string SCSS.
+ */
+function theme_foundation_pre_scss($theme) {
+    $toolbox = \theme_foundation\toolbox::get_instance();
+    return $toolbox->pre_scss('foundation');
+}
+
+/**
  * Gets the extra SCSS for the theme.
  *
  * @param theme_config $theme The theme configuration object.