+++ Changes.md
@@ -1,5 +1,8 @@
 Version Information
 ===================
+Version 3.6.0.1
+  1. Moodle 3.6rc1 beta version based on 3.5.0.6.
+
 Version 3.5.0.6
   1. Beta version.
   2. Changes for CONTRIB-7493 including a 'pre SCSS' setting.